Reference

CircoCore.Actor

CircoCore.ActorId

CircoCore.Addr

CircoCore.CircoContext

CircoCore.Pos

CircoCore.PostCode

Circo.onmigrate

CircoCore.addr

CircoCore.addr

CircoCore.become

CircoCore.box

CircoCore.box

CircoCore.die

CircoCore.getname

CircoCore.isbaseaddress

CircoCore.isnulladdr

CircoCore.onmessage

CircoCore.onspawn

CircoCore.pos

CircoCore.redirect

CircoCore.registername

CircoCore.spawn

Note

Items from CircoCore are reexported thus available directly from Circo.

CircoCore.Actor — Type

abstract type Actor{TCoreState}

Supertype of all actors.

Subtypes must be mutable and must provide a field core::TCoreState that can remain undefined after creation.

Examples

mutable struct DataHolder{TValue, TCore} <: Actor{TCore}
    value::TValue
    core::TCore
end

CircoCore.ActorId — Type

ActorId

A cluster-unique id that is randomly generated when the actor is spawned (first scheduled).

ActorId is an alias to UInt64 at the time, so it may pop up in error messages as such.

CircoCore.Addr — Type

Addr(postcode::PostCode, box::ActorId)
Addr(readable_address::String)
Addr()

The full address of an actor.

When created without arguments, it will be the null address. See isnulladdr()

If the referenced actor migrates to a different scheduler, messages sent to the old address will bounce back as RecipientMoved and the Addr must be updated manually.

Examples

Addr("192.168.1.11:24721", 0xbc6ac81fc7e4ea2)

Addr("192.168.1.11:24721/bc6ac81fc7e4ea2")

CircoCore.CircoContext — Type

CircoContext(;options...) <: AbstractContext

Store configuration, manage staging and run-time code optimizations for Circo.

CircoCore.Pos — Type

Pos(x::Real, y::Real, z::Real)
Pos(coords)

A point in the 3D "actor space".

You can access the coords by pos.x, pos.y, pos.z.

CircoCore.PostCode — Type

PostCode

A string that identifies a scheduler.

Examples

"192.168.1.11:24721"

CircoCore.addr — Method

addr(a::Actor)

Return the address of the actor.

Call this on a spawned actor to get its address. Throws UndefRefError if the actor is not spawned.

CircoCore.addr — Method

addr(entity)

Return the address of entity.

The default implementation returns the addr field, allowing you to use your own structs with such fields as message targets.

CircoCore.become — Method

become(service, old::Actor, reincarnated::Actor)

Reincarnates the old actor into new, meaning that old will die, and reincarnated will be spawned reusing the address of old.

Note: As the name suggests, become is the Circonian way of behavior change.

CircoCore.box — Method

box(a::Actor)

Return the 'P.O. box' of the spawned actor.

Call this on a spawned actor to get its id (aka box). Throws UndefRefError if the actor is not spawned.

CircoCore.box — Method

box(a::Addr)::ActorId

Return the box of the address, that is the id of the actor.

When the actor migrates, its box remains the same, only the PostCode of the address changes.

CircoCore.die — Method

die(service, me::Actor)

Unschedule the actor from its current scheduler.

CircoCore.getname — Method

function getname(service, name::String)::Union{Addr, Nothing}

Return the registered name from the scheduler-local registry, or nothing.

See also: NameQuery

CircoCore.isbaseaddress — Method

isbaseaddress(addr::Addr)::Bool

Return true if addr is a base address, meaning it references a scheduler directly.

CircoCore.isnulladdr — Method

isnulladdr(a::Addr)

Check if the given address is a null address, meaning that it points to "nowhere", messages sent to it will be dropped.

CircoCore.onmessage — Method

onmessage(me::Actor, message, service)

Hook to handle a message arriving at an actor.

Only the payload of the message is delivered, there is currently no way to access the infoton or the sender address. If you need a reply, include the sender address in the request.

Note: Do not forget to import it or use its qualified name to allow overloading!

Examples

import CircoCore.onmessage

struct TestRequest
    replyto::Addr
end

struct TestResponse end

function onmessage(me::MyActor, message::TestRequest, service)
    send(service, me, message.replyto, TestResponse())
end

CircoCore.onspawn — Method

CircoCore.onspawn(me::Actor, service)

Lifecycle callback that marks the first scheduling of the actor, called during spawning, before any onmessage.

Note: Do not forget to import it or use its qualified name to allow overloading!

Examples

import CircoCore.onspawn

funtion onspawn(me::MyActor, service)
    registername(service, "MyActor", me) # Register this actor in the local name service
end

CircoCore.pos — Method

pos(a::Actor)::Pos

return the current position of the actor.

Call this on a spawned actor to get its position. Throws UndefRefError if the actor is not spawned.

CircoCore.redirect — Method

redirect(addr::Addr, topostcode::PostCode):Addr

Create a new Addr by replacing the postcode of the given one.

CircoCore.registername — Method

registername(service, name::String, actor::Actor)

Register the given actor under the given name in the scheduler-local name registry.

Note that there is no need to unregister the name when migrating or dying

TODO implement manual and auto-unregistration

CircoCore.spawn — Method

spawn(service, actor::Actor, [pos::Pos])::Addr

Spawn the given actor on the scheduler represented by service, return the address of it.

Part of the actor API, can be called from a lifecycle callback, providing the service you got.

The onspawn callback of actor will run before this function returns.

Examples

TODO: update this sample

mutable struct ListItem{TData, TCore} <: Actor{TCore}
    data::TData
    next::Union{Nothing, Addr}
    core::TCore
    ListItem(data, core) = new{typeof(data), typeof(core)}(data, nothing, core)
end

struct Append{TData}
    value::TData
end

function CircoCore.onmessage(me::ListItem, message::Append, service)
    me.next = spawn(service, ListItem(message.value))
end

Circo.onmigrate — Method

onmigrate(me::Actor, service)

Lifecycle callback that marks a successful migration.

It is called on the target scheduler, before any messages will be delivered.

Note: Do not forget to import it or use its qualified name to allow overloading!

Examples

function Circo.onmigrate(me::MyActor, service)
    @info "Successfully migrated, registering a name on the new scheduler"
    registername(service, "MyActor", me)
end